New hasValue methods

+++ core/src/main/java/org/bouncycastle/pqc/asn1/XMSSMTPublicKey.java
@@ -1,7 +1,5 @@
 package org.bouncycastle.pqc.asn1;
 
-import java.math.BigInteger;
-
 import org.bouncycastle.asn1.ASN1EncodableVector;
 import org.bouncycastle.asn1.ASN1Integer;
 import org.bouncycastle.asn1.ASN1Object;
@@ -35,7 +33,7 @@ public class XMSSMTPublicKey
 
     private XMSSMTPublicKey(ASN1Sequence seq)
     {
-        if (!ASN1Integer.getInstance(seq.getObjectAt(0)).hasValue(BigInteger.valueOf(0)))
+        if (!ASN1Integer.getInstance(seq.getObjectAt(0)).hasValue(0))
         {
             throw new IllegalArgumentException("unknown version of sequence");
         }
